+++ lib/src/chart/line_chart/line_chart_painter.dart
@@ -1174,7 +1174,7 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       final tp = TextPainter(
           text: span,
           textAlign: tooltipItem.textAlign,
-          textDirection: TextDirection.ltr,
+          textDirection: tooltipItem.textDirection,
           textScaleFactor: holder.textScale);
       tp.layout(maxWidth: tooltipData.maxContentWidth);
       drawingTextPainters.add(tp);